requirements.txt
aiohappyeyeballs==2.4.3
aiohttp==3.10.9
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.6.0
appdirs==1.4.4
apturl==0.5.2
asttokens==3.0.0
async-timeout==4.0.3
attrs==25.1.0
autobahn==23.6.2
Automat==22.10.0
bcrypt==3.2.0
beautifulsoup4==4.13.3
blinker==1.9.0
botocore==1.35.56
Brlapi==0.8.3
build==1.2.1
buildozer==1.5.0
certifi==2023.5.7
cffi==1.15.1
chardet==4.0.0
charset-normalizer==3.1.0
click==8.1.8
colorama==0.4.4
coloredlogs==15.0.1
comm==0.2.2
command-not-found==0.3
constantly==15.1.0
cryptography==41.0.1
cupshelpers==1.0
Cython==0.29.33
-e git+https://github.com/sijumpk84/drupal_project_assistant.git@e304e87f55c065c58ff17a466af902c7f12975c9#egg=dai&subdirectory=src
dataclasses-json==0.5.14
dbus-python==1.2.18
debugpy==1.8.14
decorator==5.2.1
defer==1.0.6
distlib==0.3.8
distro==1.7.0
distro-info==1.1+ubuntu0.2
docker==5.0.3
docker-compose==1.29.2
dockerpty==0.4.1
docopt==0.6.2
docutils==0.16
duplicity==0.8.21
environs==9.5.0
exceptiongroup==1.2.2
executing==2.2.0
fastapi==0.115.0
fasteners==0.14.1
filelock==3.15.4
Flask==3.1.0
flask-cors==5.0.1
flatbuffers==24.3.25
frozenlist==1.4.1
fsspec==2024.6.1
future==0.18.2
greenlet==3.1.1
grpcio==1.63.0
gyp==0.1
h11==0.14.0
httpcore==1.0.7
httplib2==0.20.2
httpx==0.28.1
httpx-sse==0.4.0
huggingface-hub==0.24.5
humanfriendly==10.0
hyperlink==21.0.0
idna==3.4
importlib-metadata==4.6.4
incremental==22.10.0
iniconfig==2.1.0
ipykernel==6.29.5
ipython==8.35.0
isodate==0.6.1
itsdangerous==2.2.0
jedi==0.19.2
jeepney==0.7.1
Jinja2==3.1.4
jiter==0.8.2
jmespath==0.10.0
joblib==1.4.2
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==3.2.0
jupyter_client==8.6.3
jupyter_core==5.7.2
keyring==23.5.0
Kivy==2.3.0
Kivy-Garden==0.1.5
kivymd==1.2.0
langchain==0.3.25
langchain-community==0.3.24
langchain-core==0.3.59
langchain-openai==0.3.16
langchain-pinecone==0.2.6
langchain-tests==0.3.19
langchain-text-splitters==0.3.8
langsmith==0.3.42
language-selector==0.1
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
lockfile==0.12.2
logzero==1.7.0
louis==3.20.0
macaroonbakery==1.3.1
Mako==1.1.3
MarkupSafe==3.0.2
marshmallow==3.21.3
matplotlib-inline==0.1.7
meld==3.20.4
milvus-lite==2.4.8
milvus-model==0.2.3
monotonic==1.6
more-itertools==8.10.0
mpmath==1.3.0
multidict==6.1.0
mypy-extensions==1.0.0
nest-asyncio==1.6.0
netifaces==0.11.0
networkx==3.2.1
numexpr==2.10.1
numpy==1.26.4
nvidia-cublas-cu11==11.11.3.6
nvidia-cuda-cupti-cu11==11.8.87
nvidia-cuda-nvrtc-cu11==11.8.89
nvidia-cuda-runtime-cu11==11.8.89
nvidia-cudnn-cu11==9.1.0.70
nvidia-cufft-cu11==10.9.0.58
nvidia-curand-cu11==10.3.0.86
nvidia-cusolver-cu11==11.4.1.48
nvidia-cusparse-cu11==11.7.5.86
nvidia-nccl-cu11==2.20.5
nvidia-nvtx-cu11==11.8.86
oauthlib==3.2.0
olefile==0.46
onnxruntime==1.18.1
openai==1.78.1
openapi-schema-pydantic==1.2.4
orjson==3.10.18
outcome==1.3.0.post0
packaging==24.1
pandas==2.2.2
paramiko==2.9.3
parso==0.8.4
pexpect==4.8.0
Pillow==9.0.1
pinecone==6.0.2
pinecone-plugin-interface==0.0.7
platformdirs==4.2.2
pluggy==1.5.0
prompt_toolkit==3.0.51
propcache==0.2.0
protobuf==5.27.3
psutil==7.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
pyasn1==0.4.8
pycairo==1.20.1
pycparser==2.21
pycrypto==2.6.1
pycryptodome==3.21.0
pycups==2.0.1
pydantic==2.11.4
pydantic-settings==2.9.1
pydantic_core==2.33.2
pydub==0.25.1
Pygments==2.18.0
PyGObject==3.42.1
PyJWT==2.3.0
pymacaroons==0.13.0
pymilvus==2.4.5
PyNaCl==1.5.0
pyotp==2.8.0
pyparsing==3.1.0
PyPDF2==3.0.1
pyproject_hooks==1.1.0
pyRFC3339==1.1
pyrsistent==0.18.1
PySocks==1.7.1
pytest==8.3.5
pytest-asyncio==0.26.0
pytest-socket==0.7.0
python-apt==2.4.0+ubuntu4
python-dateutil==2.8.2
python-debian==0.1.43+ubuntu1.1
python-dotenv==1.0.1
pytz==2023.3
pyxdg==0.27
PyYAML==5.4.1
pyzmq==26.4.0
rdflib==6.3.2
rdflib-jsonld==0.6.2
redis==5.1.1
regex==2024.7.24
reportlab==3.6.8
requests==2.31.0
requests-toolbelt==1.0.0
roman==3.3
rsa==4.7.2
s3transfer==0.10.3
safetensors==0.4.4
scikit-learn==1.5.2
scipy==1.14.0
SecretStorage==3.3.1
selenium==4.28.1
sentence-transformers==3.1.1
sh==1.14.3
simplejson==3.19.1
six==1.16.0
smartapi-python==1.4.8
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.6
SQLAlchemy==2.0.35
stack-data==0.6.3
starlette==0.38.6
sympy==1.13.1
syrupy==4.9.1
systemd-python==234
tenacity==8.5.0
texttable==1.6.4
threadpoolctl==3.5.0
tiktoken==0.9.0
tokenizers==0.19.1
toml==0.10.2
tomli==2.0.1
torch==2.4.1+cu118
tornado==6.4.2
tqdm==4.66.5
traitlets==5.14.3
transformers==4.44.0
trio==0.28.0
trio-websocket==0.11.1
triton==3.0.0
Twisted==22.10.0
txaio==23.1.1
typing-inspect==0.9.0
typing-inspection==0.4.0
typing_extensions==4.12.2
tzdata==2024.1
ubuntu-drivers-common==0.0.0
ubuntu-pro-client==8001
ufw==0.36.1
ujson==5.10.0
unattended-upgrades==0.1
urllib3==2.2.3
usb-creator==0.3.7
uv==0.6.0
uvicorn==0.31.0
virtualenv==20.26.3
wadllib==1.3.6
wcwidth==0.2.13
webdriver-manager==4.0.2
websocket-client==1.8.0
websockets==15.0.1
Werkzeug==3.1.3
wsproto==1.2.0
xdg==5
xkit==0.0.0
yarl==1.14.0
zipp==1.0.0
zope.interface==6.0
zstandard==0.23.0
